takes txt files as args, updates readme

+++ run.py
@@ -1,24 +1,26 @@
 #!/bin/bash
-all_lyrics_file="all_lyrics.txt"
 
-# holds unshuffle lyrics
+# temp file to hold concatted, unshuffled lyrics
+all_lyrics_file="all_lyrics.tmp"
+
+if [ $# -eq 0 ]; then
+  echo "No args supplied"
+  echo "Usage: lyric-shuffler some-lyrics.txt some-more-lyrics.txt you-get-the-idea.txt"
+  exit
+fi
+
+# if file exists, remove and start afresh
 if [ -f $all_lyrics_file ]; then
   rm $all_lyrics_file
 fi
 touch $all_lyrics_file
 
-# validate songs dir
-if [ ! -d $1 ]; then
-  echo $1 is not a valid dir
-  exit
-fi
-
-# loop through *.txt files in songs dir and cat to all_lyrics file
-for f in $1/*.txt
+# loop through txt files and cat to all_lyrics.tmp file
+for f in "$@"
 do
   cat $f >> $all_lyrics_file
   echo "" >> $all_lyrics_file 
 done
 
-# shuffle up all_lyrics to stdout
+# shuffle all_lyrics, print to stdout
 ./word-shuffler.py $all_lyrics_file
